changed to template string in createGUID, and moved Utils into Models
